+++ pkgs/checks/test/extensions/async_test.dart
@@ -16,22 +16,20 @@ void main() {
   group('FutureChecks', () {
     group('completes', () {
       test('succeeds for a future that completes to a value', () async {
-        await check(_futureSuccess()).completes().which(it()..equals(42));
+        await check(_futureSuccess()).completes(it()..equals(42));
       });
       test('rejects futures which complete as errors', () async {
         await check(_futureFail()).isRejectedByAsync(
-          it()..completes().which(it()..equals(1)),
+          it()..completes(it()..equals(1)),
           actual: ['a future that completes as an error'],
           which: ['threw <UnimplementedError> at:', 'fake trace'],
         );
       });
       test('can be described', () async {
-        await check(it<Future<void>>()..completes())
-            .asyncDescription
-            .which(it()..deepEquals(['  completes to a value']));
-        await check(it<Future<int>>()..completes().which(it()..equals(42)))
-            .asyncDescription
-            .which(it()
+        await check(it<Future<void>>()..completes()).hasAsyncDescriptionWhich(
+            it()..deepEquals(['  completes to a value']));
+        await check(it<Future<int>>()..completes(it()..equals(42)))
+            .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  completes to a value that:',
                 '    equals <42>',
@@ -43,9 +41,8 @@ void main() {
       test(
           'succeeds for a future that compeletes to an error of the expected type',
           () async {
-        await check(_futureFail())
-            .throws<UnimplementedError>()
-            .which(it()..has((p0) => p0.message, 'message').isNull());
+        await check(_futureFail()).throws<UnimplementedError>(
+            it()..has((p0) => p0.message, 'message').isNull());
       });
       test('fails for futures that complete to a value', () async {
         await check(_futureSuccess()).isRejectedByAsync(
@@ -66,12 +63,10 @@ void main() {
         );
       });
       test('can be described', () async {
-        await check(it<Future<void>>()..throws())
-            .asyncDescription
-            .which(it()..deepEquals(['  completes to an error']));
+        await check(it<Future<void>>()..throws()).hasAsyncDescriptionWhich(
+            it()..deepEquals(['  completes to an error']));
         await check(it<Future<void>>()..throws<StateError>())
-            .asyncDescription
-            .which(it()
+            .hasAsyncDescriptionWhich(it()
               ..deepEquals(['  completes to an error of type StateError']));
       });
     });
@@ -122,8 +117,8 @@ fake trace''');
       });
       test('can be described', () async {
         await check(it<Future<void>>()..doesNotComplete())
-            .asyncDescription
-            .which(it()..deepEquals(['  does not complete']));
+            .hasAsyncDescriptionWhich(
+                it()..deepEquals(['  does not complete']));
       });
     });
   });
@@ -131,7 +126,7 @@ fake trace''');
   group('StreamChecks', () {
     group('emits', () {
       test('succeeds for a stream that emits a value', () async {
-        await check(_countingStream(5)).emits().which(it()..equals(0));
+        await check(_countingStream(5)).emits(it()..equals(0));
       });
       test('fails for a stream that closes without emitting', () async {
         await check(_countingStream(0)).isRejectedByAsync(
@@ -149,17 +144,15 @@ fake trace''');
       });
       test('can be described', () async {
         await check(it<StreamQueue<void>>()..emits())
-            .asyncDescription
-            .which(it()..deepEquals(['  emits a value']));
-        await check(it<StreamQueue<int>>()..emits().which(it()..equals(42)))
-            .asyncDescription
-            .which(it()
+            .hasAsyncDescriptionWhich(it()..deepEquals(['  emits a value']));
+        await check(it<StreamQueue<int>>()..emits(it()..equals(42)))
+            .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  emits a value that:',
                 '    equals <42>',
               ]));
       });
-      test('uses a transaction', () async {
+      test('does not consume error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(queue, it()..emits());
         await check(queue).emitsError();
@@ -196,16 +189,14 @@ fake trace''');
       });
       test('can be described', () async {
         await check(it<StreamQueue<void>>()..emitsError())
-            .asyncDescription
-            .which(it()..deepEquals(['  emits an error']));
+            .hasAsyncDescriptionWhich(it()..deepEquals(['  emits an error']));
         await check(it<StreamQueue<void>>()..emitsError<StateError>())
-            .asyncDescription
-            .which(it()..deepEquals(['  emits an error of type StateError']));
+            .hasAsyncDescriptionWhich(
+                it()..deepEquals(['  emits an error of type StateError']));
         await check(it<StreamQueue<void>>()
-              ..emitsError<StateError>()
-                  .which(it()..has((e) => e.message, 'message').equals('foo')))
-            .asyncDescription
-            .which(it()
+              ..emitsError<StateError>(
+                  it()..has((e) => e.message, 'message').equals('foo')))
+            .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  emits an error of type StateError that:',
                 '    has message that:',
@@ -215,7 +206,7 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1);
         await softCheckAsync<StreamQueue<int>>(queue, it()..emitsError());
-        await check(queue).emits().which((it()..equals(0)));
+        await check(queue).emits((it()..equals(0)));
       });
     });
 
@@ -234,8 +225,7 @@ fake trace''');
       });
       test('can be described', () async {
         await check(it<StreamQueue<int>>()..emitsThrough(it()..equals(42)))
-            .asyncDescription
-            .which(it()
+            .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  emits any values then emits a value that:',
                 '    equals <42>'
@@ -245,20 +235,20 @@ fake trace''');
         final queue = _countingStream(1);
         await softCheckAsync(
             queue, it<StreamQueue<int>>()..emitsThrough(it()..equals(42)));
-        check(queue).emits().which(it()..equals(0));
+        check(queue).emits(it()..equals(0));
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
         await check(queue).emitsThrough(it()..equals(1));
-        await check(queue).emits().which((it()..equals(2)));
+        await check(queue).emits((it()..equals(2)));
       });
     });
 
     group('emitsInOrder', () {
       test('succeeds for happy case', () async {
         await check(_countingStream(2)).inOrder([
-          it()..emits().which(it()..equals(0)),
-          it()..emits().which((it()..equals(1))),
+          it()..emits(it()..equals(0)),
+          it()..emits(it()..equals(1)),
           it()..isDone(),
         ]);
       });
@@ -275,7 +265,7 @@ fake trace''');
       });
       test('nestes the report for deep failures', () async {
         await check(_countingStream(2)).isRejectedByAsync(
-          it()..inOrder([it()..emits(), it()..emits().which(it()..equals(2))]),
+          it()..inOrder([it()..emits(), it()..emits(it()..equals(2))]),
           actual: ['a stream'],
           which: [
             'satisfied 1 conditions then',
@@ -289,8 +279,8 @@ fake trace''');
       });
       test('gets described with the number of conditions', () async {
         await check(it<StreamQueue<int>>()..inOrder([it(), it()]))
-            .asyncDescription
-            .which(it()..deepEquals(['  satisfies 2 conditions in order']));
+            .hasAsyncDescriptionWhich(
+                it()..deepEquals(['  satisfies 2 conditions in order']));
       });
       test('uses a transaction', () async {
         final queue = _countingStream(3);
@@ -298,21 +288,21 @@ fake trace''');
             queue,
             it()
               ..inOrder([
-                it()..emits().which(it()..equals(0)),
-                it()..emits().which(it()..equals(1)),
-                it()..emits().which(it()..equals(42)),
+                it()..emits(it()..equals(0)),
+                it()..emits(it()..equals(1)),
+                it()..emits(it()..equals(42)),
               ]));
         await check(queue).inOrder([
-          it()..emits().which(it()..equals(0)),
-          it()..emits().which(it()..equals(1)),
-          it()..emits().which(it()..equals(2)),
+          it()..emits(it()..equals(0)),
+          it()..emits(it()..equals(1)),
+          it()..emits(it()..equals(2)),
           it()..isDone(),
         ]);
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
         await check(queue).inOrder([it()..emits(), it()..emits()]);
-        await check(queue).emits().which(it()..equals(2));
+        await check(queue).emits(it()..equals(2));
       });
     });
 
@@ -331,8 +321,7 @@ fake trace''');
       });
       test('can be described', () async {
         await check(it<StreamQueue<int>>()..neverEmits(it()..equals(42)))
-            .asyncDescription
-            .which(it()
+            .hasAsyncDescriptionWhich(it()
               ..deepEquals([
                 '  never emits a value that:',
                 '    equals <42>',
@@ -343,8 +332,8 @@ fake trace''');
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..neverEmits(it()..equals(1)));
         await check(queue).inOrder([
-          it()..emits().which(it()..equals(0)),
-          it()..emits().which(it()..equals(1)),
+          it()..emits(it()..equals(0)),
+          it()..emits(it()..equals(1)),
           it()..isDone(),
         ]);
       });
@@ -364,21 +353,20 @@ fake trace''');
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(0)));
-        await check(queue).emits().which(it()..equals(1));
+        await check(queue).emits(it()..equals(1));
       });
       test('does not consume a non-matching event', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(1)));
-        await check(queue).emits().which(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('does not consume an error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmit(it()..equals(0)));
-        await check(queue)
-            .emitsError<UnimplementedError>()
-            .which(it()..has((e) => e.message, 'message').equals('Error at 1'));
+        await check(queue).emitsError<UnimplementedError>(
+            it()..has((e) => e.message, 'message').equals('Error at 1'));
       });
     });
 
@@ -397,21 +385,20 @@ fake trace''');
         final queue = _countingStream(3);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..isLessThan(2)));
-        await check(queue).emits().which(it()..equals(2));
+        await check(queue).emits(it()..equals(2));
       });
       test('consumes no events if no events match', () async {
         final queue = _countingStream(2);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..isLessThan(0)));
-        await check(queue).emits().which(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('does not consume an error', () async {
         final queue = _countingStream(1, errorAt: 0);
         await softCheckAsync<StreamQueue<int>>(
             queue, it()..mayEmitMultiple(it()..equals(0)));
-        await check(queue)
-            .emitsError<UnimplementedError>()
-            .which(it()..has((e) => e.message, 'message').equals('Error at 1'));
+        await check(queue).emitsError<UnimplementedError>(
+            it()..has((e) => e.message, 'message').equals('Error at 1'));
       });
     });
 
@@ -434,21 +421,18 @@ fake trace''');
       test('uses a transaction', () async {
         final queue = _countingStream(1);
         await softCheckAsync<StreamQueue<int>>(queue, it()..isDone());
-        await check(queue).emits().which(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('can be described', () async {
         await check(it<StreamQueue<int>>()..isDone())
-            .asyncDescription
-            .which(it()..deepEquals(['  is done']));
+            .hasAsyncDescriptionWhich(it()..deepEquals(['  is done']));
       });
     });
 
     group('emitsAnyOf', () {
       test('succeeds for a stream that matches one condition', () async {
-        await check(_countingStream(1)).anyOf([
-          it()..emits().which(it()..equals(42)),
-          it()..emits().which((it()..equals(0)))
-        ]);
+        await check(_countingStream(1)).anyOf(
+            [it()..emits(it()..equals(42)), it()..emits((it()..equals(0)))]);
       });
       test('fails for a stream that matches no conditions', () async {
         await check(_countingStream(0)).isRejectedByAsync(
@@ -472,7 +456,7 @@ fake trace''');
         await check(_countingStream(1)).isRejectedByAsync(
             it()
               ..anyOf([
-                it()..emits().which(it()..equals(42)),
+                it()..emits(it()..equals(42)),
                 it()..emitsThrough(it()..equals(10)),
               ]),
             actual: [
@@ -491,8 +475,8 @@ fake trace''');
       test('gets described with the number of conditions', () async {
         await check(
                 it<StreamQueue<int>>()..anyOf([it()..emits(), it()..emits()]))
-            .asyncDescription
-            .which(it()..deepEquals(['  satisfies any of 2 conditions']));
+            .hasAsyncDescriptionWhich(
+                it()..deepEquals(['  satisfies any of 2 conditions']));
       });
       test('uses a transaction', () async {
         final queue = _countingStream(1);
@@ -500,25 +484,19 @@ fake trace''');
             queue,
             it()
               ..anyOf([
-                it()..emits().which(it()..equals(10)),
+                it()..emits(it()..equals(10)),
                 it()..emitsThrough(it()..equals(42)),
               ]));
-        await check(queue).emits().which(it()..equals(0));
+        await check(queue).emits(it()..equals(0));
       });
       test('consumes events', () async {
         final queue = _countingStream(3);
         await check(queue).anyOf([
-          it()..emits().which(it()..equals(1)),
+          it()..emits(it()..equals(1)),
           it()..emitsThrough(it()..equals(1))
         ]);
-        await check(queue).emits().which(it()..equals(2));
+        await check(queue).emits(it()..equals(2));
       });
-    });
-  });
-
-  group('ChainAsync', () {
-    test('which', () async {
-      await check(_futureSuccess()).completes().which(it()..equals(42));
     });
   });